requirements-dev.txt
# Dependências de desenvolvimento

# Inclui todas as dependências principais
-r requirements.txt

# Ferramentas de desenvolvimento e teste
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
coverage==7.6.9

# Formatação e linting
black==25.1.0
flake8==7.1.1
isort==5.13.2

# Debug e performance
django-debug-toolbar==4.4.8
django-extensions==3.2.3

# Documentação
sphinx==8.1.3
sphinx-rtd-theme==3.0.2